perf: have DeleteNamespaces skip rel deletes when invoked from WriteSchema (#2697)

+++ internal/datastore/memdb/readwrite.go
@@ -298,7 +298,11 @@ func (rwt *memdbReadWriteTx) WriteNamespaces(_ context.Context, newConfigs ...*c
 	return nil
 }
 
-func (rwt *memdbReadWriteTx) DeleteNamespaces(_ context.Context, nsNames ...string) error {
+func (rwt *memdbReadWriteTx) DeleteNamespaces(_ context.Context, nsNames []string, delOption datastore.DeleteNamespacesRelationshipsOption) error {
+	if len(nsNames) == 0 {
+		return nil
+	}
+
 	rwt.mustLock()
 	defer rwt.Unlock()
 
@@ -321,11 +325,13 @@ func (rwt *memdbReadWriteTx) DeleteNamespaces(_ context.Context, nsNames ...stri
 			return err
 		}
 
-		// Delete the relationships from the namespace
-		if _, _, err := rwt.deleteWithLock(tx, &v1.RelationshipFilter{
-			ResourceType: nsName,
-		}, 0); err != nil {
-			return fmt.Errorf("unable to delete relationships from deleted namespace: %w", err)
+		if delOption == datastore.DeleteNamespacesAndRelationships {
+			// Delete the relationships from the namespace
+			if _, _, err := rwt.deleteWithLock(tx, &v1.RelationshipFilter{
+				ResourceType: nsName,
+			}, 0); err != nil {
+				return fmt.Errorf("unable to delete relationships from deleted namespace: %w", err)
+			}
 		}
 	}
